# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created. Password hashes are never returned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 201
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserDto"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "409": {
            "description": "Email already exists, or a global role slug is unknown/inactive."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a platform user (optionally with global role slugs).",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque nextCursor from the previous page (base64url of the last item id).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (default 20, max 100).",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cursor page of users (items + nextCursor).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserDto"
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List platform users (cursor pagination).",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User id (uuid).",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserDto"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch one user by id.",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User id (uuid).",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserDto"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "404": {
            "description": "User not found."
          },
          "409": {
            "description": "Cannot deactivate the last active super admin (409)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user name/active flag/password (deactivation revokes sessions).",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User id (uuid).",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted; data is null.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "nullable": true,
                      "example": null
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "404": {
            "description": "User not found."
          },
          "409": {
            "description": "Cannot delete the last active super admin."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a user (self-lockout protected).",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}/roles": {
      "put": {
        "operationId": "UsersController_setRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User id (uuid).",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetUserRolesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user; outstanding tokens are invalidated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserDto"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "404": {
            "description": "User not found."
          },
          "409": {
            "description": "Cannot remove the last active super admin, or a slug is unknown/inactive."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace the user global role set (bumps authVersion, revokes sessions).",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bus Fleet API",
    "description": "Multi-tenant group bus transportation platform REST API.\n\n- **Authentication is owned by this API**: HS256 JWTs carry a custom `app_role` claim; refresh tokens rotate on every use and are stored hashed at rest.\n- **Every response is wrapped in a `{ statusCode, data }` envelope**; the `data` schema is shown inside each documented response.\n- **Fleet isolation is enforced by PostgreSQL RLS** as the final security boundary; cross-tenant lookups are indistinguishable from missing rows and return 404.\n- `fleetId` path parameters are **selectors, never proof of authorization** — access is verified against the caller's ACTIVE `FleetMember` row on every request.\n- Roles and permissions are **database-driven** (no hardcoded enums); only `super_admin` is a predefined system role. Platform routes (roles, permissions, users, fleets admin, audit-logs) are privileged: super admin only and audited.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "example": "owner@fleet.com",
            "format": "email"
          },
          "name": {
            "type": "string",
            "example": "Waleed"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "authVersion": {
            "type": "number",
            "example": 1,
            "description": "Bumped on security-sensitive changes; stale tokens rejected."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "isActive",
          "authVersion",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "owner@fleet.com",
            "format": "email",
            "maxLength": 255,
            "description": "Stored lowercase; duplicates → 409."
          },
          "password": {
            "type": "string",
            "example": "Passw0rd!123",
            "format": "password",
            "minLength": 8,
            "maxLength": 128
          },
          "name": {
            "type": "string",
            "example": "Waleed",
            "maxLength": 255
          },
          "globalRoleSlugs": {
            "example": [
              "fleet-owner"
            ],
            "description": "Global role slugs assigned at creation (unknown/inactive slugs → 409).",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Waleed",
            "maxLength": 255
          },
          "isActive": {
            "type": "boolean",
            "example": false,
            "description": "Deactivating bumps authVersion and revokes sessions."
          },
          "password": {
            "type": "string",
            "example": "Passw0rd!456",
            "format": "password",
            "minLength": 8,
            "maxLength": 128
          }
        }
      },
      "SetUserRolesDto": {
        "type": "object",
        "properties": {
          "roleSlugs": {
            "example": [
              "fleet-owner"
            ],
            "description": "Full replacement set of global role slugs. Removing the last active super admin → 409.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "roleSlugs"
        ]
      },
...
    }
  }
}